requirements.txt
discord.py==1.7.1
python-dotenv==0.17.0
requests==2.22.0
urllib3==1.25.8
